// apps/intelli-law/lib/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      documents: {
        Row: {
          account_id: string
          content: string
          created_at: string
          id: string
          summary: string
          title: string
        }
        Insert: {
          account_id: string
          content: string
          created_at?: string
          id?: string
          summary?: string
          title: string
        }
        Update: {
          account_id?: string
          content?: string
          created_at?: string
          id?: string
          summary?: string
          title?: string
        }
        Relationships: [
          {
            foreignKeyName: "documents_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "accounts"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "documents_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "user_account_workspace"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "documents_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "user_accounts"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
// ...
}